+++ app/javascript/sdk/specs/cookieHelpers.spec.js
@@ -1,17 +1,9 @@
 import Cookies from 'js-cookie';
 import {
-  getUserCookieName,
   getUserString,
   hasUserKeys,
   setCookieWithDomain,
 } from '../cookieHelpers';
-
-describe('#getUserCookieName', () => {
-  it('returns correct cookie name', () => {
-    global.$chatwoot = { websiteToken: '123456' };
-    expect(getUserCookieName()).toBe('cw_user_123456');
-  });
-});
 
 describe('#getUserString', () => {
   it('returns correct user string', () => {
@@ -22,11 +14,12 @@ describe('#getUserString', () => {
           email: 'pranav@example.com',
           avatar_url: 'https://images.chatwoot.com/placeholder',
           identifier_hash: '12345',
+          phone_number: '+15551234',
         },
         identifier: '12345',
       })
     ).toBe(
-      'avatar_urlhttps://images.chatwoot.com/placeholderemailpranav@example.comnamePranavidentifier_hash12345identifier12345'
+      '[["avatar_url","https://images.chatwoot.com/placeholder"],["email","pranav@example.com"],["name","Pranav"],["identifier_hash","12345"],["phone_number","+15551234"],["identifier","12345"]]'
     );
 
     expect(
@@ -37,7 +30,7 @@ describe('#getUserString', () => {
         },
       })
     ).toBe(
-      'avatar_urlhttps://images.chatwoot.com/placeholderemailpranav@example.comnameidentifier_hashidentifier'
+      '[["avatar_url","https://images.chatwoot.com/placeholder"],["email","pranav@example.com"],["name",""],["identifier_hash",""],["identifier",""]]'
     );
   });
 });